backend/requirements.txt
pymongo~=4.16.0
fastapi~=0.104.1
uvicorn[standard]~=0.24.0
python-dotenv~=1.0.0